services/catalog-api/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-filter==24.3
gunicorn==21.2.0
pika==1.3.2
Pillow